Prime-web is building after updating dependencies. Including package-lock.json.
Updated package.json, mostly removing caret operators which lead to updated packages (which used Angular 5).

Suppressed an unavoidable TSC warning. AngularCLI 1.4.4 doesn't officially support TSC 2.4.0, but in practice we're fine. We use some features of TSC 2.4.0, mostly to do with enums, and it's not as simple as updating package versions.